🚚  Group most of the types together, in types/index.ts
