+++ src/commands/basic/help.ts
@@ -2,9 +2,9 @@ import { Command } from 'discord-akairo';
 import { MessageEmbed } from 'discord.js';
 import { help as config } from '../../../config/commands/basic';
 import settings from '../../../config/settings';
+import { Rules } from '../../types';
+import type { GuildMessage } from '../../types';
 import type { HelpCommandArguments } from '../../types/CommandArguments';
-import Rules from '../../types/rules';
-import type { GuildMessage } from '../../types/utils';
 import { capitalize } from '../../utils';
 
 class HelpCommand extends Command {
