+++ src/commands/basic/links.ts
@@ -2,8 +2,8 @@ import { Argument, Command } from 'discord-akairo';
 import { MessageEmbed } from 'discord.js';
 import { links as config } from '../../../config/commands/basic';
 import settings from '../../../config/settings';
+import type { GuildMessage } from '../../types';
 import type { LinksCommandArguments } from '../../types/CommandArguments';
-import type { GuildMessage } from '../../types/utils';
 
 const reactions = ['⏮', '◀', '🇽', '▶', '⏭'];
 const maxPage = 5;
