+++ src/commands/basic/ping.ts
@@ -2,9 +2,9 @@ import { Command } from 'discord-akairo';
 import { MessageEmbed } from 'discord.js';
 import { ping as config } from '../../../config/commands/basic';
 import settings from '../../../config/settings';
+import { Rules } from '../../types';
+import type { GuildMessage } from '../../types';
 import type { PingCommandArguments } from '../../types/CommandArguments';
-import Rules from '../../types/rules';
-import type { GuildMessage } from '../../types/utils';
 
 class PingCommand extends Command {
   constructor() {
