+++ src/commands/moderation/history.ts
@@ -1,13 +1,14 @@
 import { oneLine, stripIndent } from 'common-tags';
 import { Argument, Command } from 'discord-akairo';
+import { GuildMember } from 'discord.js';
 import moment from 'moment';
 import { history as config } from '../../../config/commands/moderation';
 import messages from '../../../config/messages';
 import settings from '../../../config/settings';
 import Sanction from '../../models/sanction';
+import { SanctionsUpdates, SanctionCreations } from '../../types';
+import type { GuildMessage } from '../../types';
 import type { HistoryCommandArgument } from '../../types/CommandArguments';
-import { SanctionsUpdates, SanctionCreations } from '../../types/sanctionsTypes';
-import type { GuildMessage } from '../../types/utils';
 import { noop, splitText, toHumanDuration } from '../../utils';
 
 class HistoryCommand extends Command {
@@ -31,6 +32,12 @@ class HistoryCommand extends Command {
 
   public async exec(message: GuildMessage, args: HistoryCommandArgument): Promise<void> {
     const memberId = typeof args.member === 'string' ? args.member : args.member.id;
+    const memberName = typeof args.member === 'string'
+      ? args.member
+      : (args.member instanceof GuildMember
+        ? args.member.displayName
+        : args.member.username
+      );
 
     const sanctions = await Sanction.find({ memberId });
     if (sanctions.length === 0) {
@@ -46,8 +53,7 @@ class HistoryCommand extends Command {
     };
 
     let privateHistory = config.messages.title
-      // @ts-expect-error
-      .replace('{NAME}', args.member?.displayName || args.member?.username || args.member)
+      .replace('{NAME}', memberName)
       .replace('{COUNT}', sanctions.length.toString());
 
     privateHistory += config.messages.overview
