+++ src/commands/moderation/kick.ts
@@ -4,9 +4,9 @@ import messages from '../../../config/messages';
 import ModerationData from '../../moderation/ModerationData';
 import KickAction from '../../moderation/actions/KickAction';
 import Logger from '../../structures/Logger';
+import { SanctionTypes } from '../../types';
+import type { GuildMessage } from '../../types';
 import type { KickCommandArgument } from '../../types/CommandArguments';
-import { SanctionTypes } from '../../types/sanctionsTypes';
-import type { GuildMessage } from '../../types/utils';
 import { noop } from '../../utils';
 
 class KickCommand extends Command {
