+++ src/commands/moderation/mute.ts
@@ -6,9 +6,9 @@ import ModerationData from '../../moderation/ModerationData';
 import ModerationHelper from '../../moderation/ModerationHelper';
 import MuteAction from '../../moderation/actions/MuteAction';
 import Logger from '../../structures/Logger';
+import { SanctionTypes } from '../../types';
+import type { GuildMessage } from '../../types';
 import type { MuteCommandArgument } from '../../types/CommandArguments';
-import { SanctionTypes } from '../../types/sanctionsTypes';
-import type { GuildMessage } from '../../types/utils';
 import { noop } from '../../utils';
 
 class MuteCommand extends Command {
