+++ src/moderation/actions/MuteAction.ts
@@ -7,7 +7,7 @@ import {
 import settings from '../../../config/settings';
 import ConvictedUser from '../../models/convictedUser';
 import Sanction from '../../models/sanction';
-import { SanctionsUpdates } from '../../types/sanctionsTypes';
+import { SanctionsUpdates } from '../../types';
 import ModerationError from '../ModerationError';
 import ModerationAction from './ModerationAction';
