+++ src/moderation/actions/RemoveWarnAction.ts
@@ -1,7 +1,7 @@
 import { User } from 'discord.js';
 import ConvictedUser from '../../models/convictedUser';
 import Sanction from '../../models/sanction';
-import { SanctionCreations, SanctionsUpdates } from '../../types/sanctionsTypes';
+import { SanctionCreations, SanctionsUpdates } from '../../types';
 import ModerationError from '../ModerationError';
 import ModerationAction from './ModerationAction';
