+++ src/moderation/actions/UnmuteAction.ts
@@ -2,7 +2,7 @@ import { GuildMember, User, Permissions } from 'discord.js';
 import settings from '../../../config/settings';
 import ConvictedUser from '../../models/convictedUser';
 import Sanction from '../../models/sanction';
-import { SanctionsUpdates } from '../../types/sanctionsTypes';
+import { SanctionsUpdates } from '../../types';
 import ModerationError from '../ModerationError';
 import ModerationAction from './ModerationAction';
