--- src/types/rules.ts
@@ -1,7 +1,0 @@
-enum Rules {
-  OnlyBotChannel,
-  NoHelpChannel,
-  OnlyHelpChannel,
-}
-
-export default Rules;
